Refactor package "command"
