+++ user_management_service/app/commands/vault/model/AppRoleCredentials.scala
@@ -1,4 +1,4 @@
-package command.model
+package commands.vault.model
 
 import play.api.libs.json.Json
